+++ src/kard/organizations/content_strategies/types/content_strategy_attributes.py
@@ -7,7 +7,7 @@ import typing_extensions
 from ....commons.types.category_option import CategoryOption
 from ....core.pydantic_utilities import IS_PYDANTIC_V2, UniversalBaseModel
 from ....core.serialization import FieldMetadata
-from .content_strategy_filter import ContentStrategyFilter
+from .content_strategy_sort import ContentStrategySort
 
 
 class ContentStrategyAttributes(UniversalBaseModel):
@@ -25,9 +25,9 @@ class ContentStrategyAttributes(UniversalBaseModel):
         FieldMetadata(alias="organizationId"),
         pydantic.Field(alias="organizationId", description="ID of the organization this content strategy belongs to"),
     ]
-    filter: typing.Optional[ContentStrategyFilter] = pydantic.Field(default=None)
+    sort: typing.Optional[ContentStrategySort] = pydantic.Field(default=None)
     """
-    Filter applied when selecting offers for the strategy
+    Sort applied when selecting offers for the strategy
     """
 
     categories: typing.List[CategoryOption] = pydantic.Field()